# Fix: Lazily create token value object when requesting it