This PR

* [x] improves performance by lazily creating the `Token` value object when requesting it, rather than iterating over the entire sequence

### Before

```
+-------------+----------------------------+--------+--------+------+-----+------------+-------------+-------------+-------------+-------------+---------+--------+----------+
| benchmark   | subject                    | groups | params | revs | its | mem_peak   | best        | mean        | mode        | worst       | stdev   | rstdev | diff     |
+-------------+----------------------------+--------+--------+------+-----+------------+-------------+-------------+-------------+-------------+---------+--------+----------+
| ClassyBench | benchTokenGetAllFromSource |        | []     | 100  | 1   | 4,700,808b | 1,233.620μs | 1,233.620μs | 1,233.620μs | 1,233.620μs | 0.000μs | 0.00%  | 0.00%    |
| ClassyBench | benchSequenceFromSource    |        | []     | 100  | 1   | 7,321,488b | 6,743.240μs | 6,743.240μs | 6,743.240μs | 6,743.240μs | 0.000μs | 0.00%  | +446.62% |
+-------------+----------------------------+--------+--------+------+-----+------------+-------------+-------------+-------------+-------------+---------+--------+----------+
```

### After

```
+-------------+----------------------------+--------+--------+------+-----+------------+-------------+-------------+-------------+-------------+---------+--------+---------+
| benchmark   | subject                    | groups | params | revs | its | mem_peak   | best        | mean        | mode        | worst       | stdev   | rstdev | diff    |
+-------------+----------------------------+--------+--------+------+-----+------------+-------------+-------------+-------------+-------------+---------+--------+---------+
| ClassyBench | benchTokenGetAllFromSource |        | []     | 100  | 1   | 4,700,808b | 1,225.360μs | 1,225.360μs | 1,225.360μs | 1,225.360μs | 0.000μs | 0.00%  | 0.00%   |
| ClassyBench | benchSequenceFromSource    |        | []     | 100  | 1   | 5,171,760b | 1,524.940μs | 1,524.940μs | 1,524.940μs | 1,524.940μs | 0.000μs | 0.00%  | +24.45% |
+-------------+----------------------------+--------+--------+------+-----+------------+-------------+-------------+-------------+-------------+---------+--------+---------+
```